[FIRRTL] Convert SerializedViewAnnotation str, NFC
Change the internal name used to store a
"GrandCentralView$SerializeViewAnnotation" to just "ViewAnnotation".
The former is an artifact of the Grand Central implementation where a
special "Serialized" (JSON String encoded JSON) version of a
ViewAnnotation is passed into the FIRRTL compiler because a Scala-based
FIRRTL compiler cannot serialize classes it doesn't know about.  Because
we do this deserialization during parsing, it's cleaner to just store
this as a "ViewAnnotation" rather than to cling to an unnecessarily long
name which doesn't reflect the fact that we've deserialized the
annotation already.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>