[CMake] Set CXX_VISIBILITY_PRESET to "default" for some libs, NFC.
This should be in line with the CMake default, but in an upcoming
change, I would like to update the default we build with to
"hidden". This will let us change the default without worrying about
these specific shared libraries, or annotating each function in the
libraries with a visibility attribute.