Revert "[FIRRTL][LowerToHW] Handle 0-bit mux sel (#1349)"
Revert e1f33d13dddd54ced74ae9b417510c3e12164386.

This reverts changes that allowed a FIRRTL mux with a 0-bit condition to
both be constructed and folded to the false condition.  The consensus is
that this is something that should _not_ be allowed in the FIRRTL spec
and should be an error.

Fixes #1471.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>