[SV][ExtractTestCode] Extract test code output paths from the annotation (#1459)
Process `ExtractTestCode` annotations, for the directory path and bind file

The annotations "ExtractAssertionsAnnotation, ExtractAssumptionsAnnotation
 and ExtractCoverageAnnotation" are used to used to specify the directory
 for extracting the test code with Assert/Assume/Cover statements and
also the file for the corresponding bind statements.
Example annotation:
{class = "sifive.enterprise.firrtl.ExtractAssumptionsAnnotation",
directory = "/dir1/dir2/dir3/assumptions",
filename = "/dir1/dir2/dir3/assumptions/assumptions.bindings.sv"}


The annotations will be dropped after the LowerToHW, but the extraction
of assert/assume/cover to the external module happens at a later pass,
SVExtractTestCode. The directory information from the annotations will
 be saved in the IR as "output_file" attribute on the corresponding
assert/assume/cover statement.
Then, during the SVExtractTestCode the simulation constructs
assert/assume/cover will be extracted to a separate module.
The output_file attribute will be recorded on the module
definition. This ensures that the simulation constructs are
extracted to the appropriate directories and the bind op 
to the appropriate file.
Changes::
1. Update `LowerToHW` to process remaining annotations before
warning on them.
2. Add `hw::OutputFileAttr` from the corresponding annotation
to `assume, cover, assert` operations.
3. Update `SVExtractTestCode` to add the `hw::OutputFileAttr`
attribute to the extracted module and bind ops.