[FIRRTL] Stop or/xor invalid folds if types differ
Stop folding or/xor with invalid operands when the types of the
non-invalid operand doesn't match the result type.  This covers the case
where or and xor can take SInt operands and return a UInt result.

Include a temporary test that checks that this behavior works.

This behavior is expected to be changed in the future via a
canonicalizer that will do the cast as necessary.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>