[CMake] Remove setting CMAKE_CXX_VISIBILITY_PRESET. (#1544)
Previously, this would always set visibility=hidden for static,
unified builds. This configuration is great for packaging Python
artifacts, but setting visibility=hidden doesn't always make sense.

Instead, this removes that, leaving visibility=default as the default,
and allowing users or build bots to override it on the cmake command
line when necessary.

The sanity check is left in place to prevent visibility=hidden with a
BUILD_SHARED_LIBS=ON build. MLIR needs significant changes to support
such a configuration.