[HW] Support TypeAliasType in return type for aggregate create ops. (#1521)
This was always intended to be supported, but was removed in
6e6ddfb. This adds it back, and updates the parsers for the relevant
ops to actually support type aliases (they previously didn't).

In order to help avoid the need to cast away aliases everywhere, a new
base class is added to all aggregate types that support aliases. This
class can represent a type alias or the aggregate type. It implicitly
converts to the aggregate type, which helps avoid explicit casts.