[Comb] Enhance the icmp/compare folder to use a "computeMaskedBits" analysis.
This generalizes the analysis beyond literal constants to handle some simple
expressions that generate known bits, starting with 'and' and 'or' expressions.
Given a value with some known and some unknown bits, we extract out the portion
of the expression that is unknown and compare just it instead of the whole
value.  This composes very naturally with our existing 'truncate folding'
optimizations.

This turns the example in Issue #1560 from:

```
  wire [37:0] _tmp = (auto_in_aw_bits_addr ^ 38'h40000000) & 38'h3FE0000000;
  assign a = _tmp == 38'h0;
```

into:

```
  assign a = (auto_in_aw_bits_addr[37:29] ^ 9'h2) == 9'h0;
```

Which is a nice step forward and enables xor folding.  However, I realize
now that this whole thing can be far more general than icmp of concats.
I will take care of this in a follow-up patch.