[LowerToHW] Fix issue #1601, changing the uniquing algorithm for always blocks.
When emitting two always/initial/... blocks with the same conditions, leave the
resultant block at the location of the last use, instead of at the location of
the first use.  This ensures that the values used by that block are defined above
it.

This isn't necessary by the semantics of our IR, since hw.module is a graph region,
but leads to better generated verilog and puts less pressure on the cleanup passes.