[ExportVerilog] Refactor out of line expression emission, NFC.
We previously had a bunch of special cases for constants that
worked but was a bit fragile. The issue is that there was coupling
between stuff that "knew" that out-of-line constants in procedural
regions are emitted in the declaration.

Clean this up a bit by adding a expressionsEmittedIntoDecl set,
which makes this more explicit and easier to generalize.

NFC.