[LowerCHIRRTL] Fix inifinite loop in memory direction inference
When infering the direction of a CHIRRTL memory, we look through
subindexes, subfields, and subaccesses looking for connect ops which
could be reading or writing to the memory.  The code to look through
subaccess was broken, leading to an infinite loop. This properly
recurses on the results of the subaccess op instead of the input.