[HW] Introduce a new hw::SymbolCache class to accelerate symbol lookups.
The HW/SV dialects have a bunch of symbols used to map instances, modules,
interfaces etc, and corresponding "getReferencedXXX" methods to resolve
them.  Each of these resolutions is really slow - scanning huge amounts of
the IR to resolve them, so we should have a way to shortcut that when a
client has done a prepass.  The SymbolCache class allows making these faster.

This patch introduces the new functionality, but no clients of it.