[VerilogEmitter] Fix a race with sv.binds inside of hw.modules.
It turns out that sv.bind can exist inside of a hw.module as well
as at the top level.  These cause non-local references so we need
to make sure not to emit such modules concurrently with the modules
being emitted.

There is no good test for this, I noticed it due to a transient
failure of the existing ExportVerilog/sv-dialect.mlir test.