[FIRRTL] Fix Subfield/Subindex DontTouch Parsing
Fix a bug in the parser where a DontTouchAnnotation on a subfield would
not cause names to be preserved.  Use the simplest solution here of
using the hasDontTouch method of AnnotationSet as opposed to the older
version of the method of the same name in the parser.  Remove the old
hasDontTouch method in the parser (since this is now unused).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>