[Calyx] Parse attributes for BlockArguments and emit Calyx attributes. (#1727)
This PR is a collective effort to add Calyx attributes and emit them in the native compiler. This requires the addition of 
attributes for BlockArguments. The second part of the effort is proper emission of attributes that are found in the native 
compiler. Lastly, we add an optional attribute dictionary to the ComponentOp.