[FIRRTL] GCT Views: Scatter, Use Bundle Name
Fix a bug in Grand Central Views/Interfaces where an interface instance
name was improperly derived from the name of the interface.  Change
scattering to include the instance name (the "name" field) with the
AugmentedBundleType.  After this, the interface name comes first and the
instance name comes second, e.g., "Foo foo()".  Rework the Grand
Central-internal VerbatimType to better support this.

Update tests to exercise this change.  Tests use a convention where the
instance name begins with an uppercase letter and the interface name
begins with an lowercase letter.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>