[Calyx] Verify CombGroupOp is combinational and certain primitive ports are driven together. (#1764)
This PR focuses on (2) objectives. First, verifying that combinational groups are indeed combinational. 
Combinational components are not supported yet, so component instances are always assumed to be 
combinational for now. Second, we also verify that certain ports are written or read together. For example, 
it does not make sense to write to a register without write_en being high. To aid in this, I added helper 
functions to access certain ports on primitive operations.