[Calyx] Verify that sources of group ports and assignments are not logically complex. (#1774)
As discussed in #1769, some verifiers should be added to avoid giving the sources of group ports / assignments 
logically complex expressions. Also adds a verifier to ensure that GroupOp isn't being used as a CombGroupOp, and 
makes the necessary refactoring. This will be removed when native Calyx supports complex arguments for sources.