[Calyx] Add combinational trait. (#1781)
This adds a combinational trait to Calyx operations. We can throw this on all non-memory operations for now, 
since they are all combinational. Eventually we'll need to split this off. Closes #1739. Also fixes a bug where it 
was considering a memory load as a store, since the address ports are used when both storing and loading to 
a MemoryOp. A test is added to verify this no longer occurs.