[FIRRTL][HW] Add Support for "Port Order" Write-Under-Write to HWMemSimImpl, Use for FIRRTL Lowering (#1805)
Change the way that FIRRTL memories are lowered to HW memories to
propagate the write-under-write behavior.  Modify HWMemSimImpl (which
generate a simulation model of a memory) to respect a new
write-under-write attribute.

This new write-under-write attribute can either be "undefined"
indicating that the resulting simulation memory should put each write
port in an independent always block, regardless of whether or not they
are on the same clock.  If the write-under-write behavior is set to
"port order" then write ports that are trivially driven by the same
clock are put in the same always block and ordered sequentially by port
order.  I.e., later ports will override earlier ports if they both write
to the same address.

This behavior is added because the Scala FIRRTL Compiler (SFC) will
lower write ports in this way and a deviation from it is a formal
equivalence mismatch.

Add tests of "port order" write-under-write behavior for simulation
memory generation.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>