[FIRRTL][Inliner] Don't delete unknown ops when deleting dead modules (#1826)
The inliner removes any module which is left uninstantiated after all
inlining is complete.  It was mistakenly deleting non-module operations
because there was no instance of them.  For example, an `sv.verbatim`
operation would be deleted since it is not the sort of thing that could
be instantiated.  This adds a check that the operation it is considering
for deletion is in fact a module operation before removing it.