[FIRRTL] Add EmitMetada pass with RetimeModulesAnnotation support
This adds a new `EmitMetadata` pass to FIRRTL which a dumping ground for
any simple metadata generation passes that don't make sense to implement
as a standalone pass.  The first pass implemented is `RetimeModules`
which just collects the name of any annotated module and emits it into a
JSON array.