[Calyx] [SCFToCalyx] Add EnableOp builder which doesn't require the `compiledGroups` attribute. (#1854)
The compiledGroups attribute is only used during the (currently incomplete) lowering process of Calyx IR 
to signify which groups are attributed to the calyx.enable during compilation. This maintains the invariant 
that all Group symbols are referenced throughout the life time of a Calyx program. This PR adds another 
builder to skip initializing it. Closes #1844.