[SCFToCalyx] Support memref operations [12/12] (#1863)
This commit adds support for memory accessing operations.

All index types are converted to a fixed-width integer. This is necessary due to the lack of a bitwidth inference pass. Upon an index-typed value being used as a memory address input, the address value is truncated to the width of the memory port.