[FIRRTL][ExpandWhens] Properly track all outer scope value (#1882)
When ExpandWhens sees the the following code, it *should* produce the
following:

```firrtl
out <= in0
when p0 :
  when p1 :
    out <= in1
```

```firrtl
out <= mux(p0, mux(p1, in1, in0), in0)
```

There was a mistake in the code where it would only track previously
driven values from one level of outer scope.  This means that when
producing the inner mux, `mux(p1, in1, in0)`, it never found that `in0`
was previously driven to `out`, and never formed a mux with the `in1`
connection:

```firrtl
out <= mux(p0, in1, in0)
```

To solve this, we had to make sure that the all scopes stored previously
driven values in the same hashtable.  This allowed the code to find the
previously driven value from all surrounding scopes.

At the same time, we needed to be able to remove all elements added in
scope from the global hashtable.  For example, after processing a `then`
block we need to reuse the global hashtable for the `else` block without
the modifcations of the `then` block.

To avoid this problem, we could create a copy of the hashtable before
recursing on WhenOps.  Instead of this, this adds a HashTableStack that
allows us to quickly pop off nested hashtables to reset the state. We
don't use a ScopedHashTable so that we can pop a scope and keep the
scope information for merging `then` and `else` scopes into the parent
scope.

Fixes #1878