[HandshakeToFIRRTL] Rework submodule name generation (#1876)
This commit introduces modifications to how FIRRTL submodule names are generated. Most importantly, submodules are now generated based on both their input- and output types.

A downside of this approach is that components with a large number of i/o ports will have rather verbose names. If we see this as a problem, we can solve this with a few special cases. However, the proposed approach in this commit attempts to provide a general naming approach and avoid special cases.