[FIRRTL][LowerToHW] Rename and reorder FIRRTL memory ports (#1899)
Rename and reorder the memory ports when lowering FIRRTL memories. This PR makes sure
the ports match with the Scala FIRRTL Compiler, to be compatible with other
memory generators. 
Also update the `HWMemSimImpl` to respect the proper port numbers.
This PR fixes the issue with https://github.com/llvm/circt/pull/1891, which was reverted.