[Calyx] Add enhancements to common tail elimination for `IfOp`. (#1895)
This adds a pattern for when then and else regions are both ParOps. It also takes into consideration 
the parent region of the IfOp, since we don't want to unintentionally change the behavior when pulling 
out EnableOps. I also do not touch anything that's not immediately within the given ParOp's body; 
doing so may also change the behavior of the program. The case where the then and else regions are 
different, e.g. SeqOp and ParOp, should never pull out an EnableOp; it will always produce worse code.