[FIRRTL] Add first implementation of EmitSignalMappings (#1665)
Add a first implementation of Grand Central's `EmitSignalMappings` pass,
which populates a module in a circuit with `force` statements and
cross-module references that access things in a second circuit. This 
requires a lot of polishing and ironing out of kinks and the design in 
general later, but this is a first bare-bones implementation in the 
spirit of the other Grand Central passes (use verbatim SV to get 
something going, even if it's brittle and needs some cleaning up later).

To avoid having to deal with the issue that `EmitSignalMappings` 
actually operates on two circuits in parallel, the current 
implementation only operates on the circuit where the `force` statements
are to be inserted, and uses some text manipulation to guess what name 
the forced/probed things in the remote circuit will end up having in the 
output Verilog. This will need proper XMRs in the future, but works as a 
first shot.

Co-authored-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>