[SCFToCalyx] Add support for `memref` function arguments (#1897)
This commit adds support for `memref` function arguments. This should greatly expand the number of meaningful accelerators which we are able to generate through `SCFToCalyx`.
The signature of the top-level component is rewritten to include in- and output ports expected for the memory. These ports are written into a new `CalyxMemoryPorts` structure. This, alongside rewriting memory getter/setters to use `CalyxMemoryInterface` allows the external memories to be used like `calyx::MemoryOp`s within the remainder of the code.

The implementation assumes identical semantics for external memories as for Calyx memories, namely that reads are combinational.