[Calyx] Add canonicalization for empty bodies. (#1907)
This adds canonicalization patterns for empty IfOp, WhileOp, SeqOp, and ParOp bodies. For the former two, 
we also check to see if the (optional) Group and conditional are used. if not, we remove them as well. This also 
loosens the verification restrictions for empty bodies in a WhileOp and IfOp. IMO, these are still valid with an 
empty body.