[FIRRTL][InferResets] Use instance graph
InferResets is using a `unique_ptr<SymbolTable>()` which was improperly
free'd causing a memory leak.  The `release` call just releases ownship
from the `unique_ptr` but does not free the memory.

This switches the pass to use an `InstanceGraph`.  Since this pass does
not change the hiearchy it can always mark this analysis as preserved.
This has the benefit of sharing the analysis between passes instead of
calculating a new `SymbolTable`.