[PyCDE] Move from returning output values to setting them (#1919)
Adds a setter on the generator argument which sets the value of the output
port. Also provides a `set_all_ports` method for programatic access and easy
legacy conversion.

There are two advantages of this: (1) Symbolic port names vs. string port
names. (2) More importantly, exceptions are raised when the code tries to set an
output port value, which provides a traceback to the relevant line.