[Calyx] [SCFToCalyx] Change Cell instance name from StringAttr to FlatSymbolRefAttr. (#1924)
Changes a cell's instanceName argument from StringAttr to FlatSymbolRefAttr. This will be necessary for the 
addition of an InvokeOp, which needs to verify that the instance being invoked actually exists. Since primitives 
(namely multi-cycle ones) may also be invokeable, this is required for all cells. This also verifies there is no name
shadowing with symbol names for components and cells.