[FIRRTL] Add scatterOMIR function, NFC
Add a function that can be used to "scatter" and mutate Object Model 2.0
IR JSON.  This is primarily doing two things: (1) OMIR is updated to
make it less stringly-typed by breaking apart strings encoding a
type/value into explicit "tpe" and "value" fields and (2) "value"
members that are targets generate trackers in the circuit.

This function is not used in this patch.  This is part of incremental OM
2.0 support.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>