[FIRRTL] Simplify OMIR JSON Parsing (do no work)
Change Object Model 2.0 JSON parsing to do nothing clever and just parse
the OMIR as JSON into an OMIRAnnotation.  This is needed as any
"cleverness" needs to happen for both OMIR files and for parsed
OMIRAnnotations.  Specifically, this removes conversion of source
locators to LocationAttrs and changing the fields OMNode member from an
array to a dictionary.  This will be moved into OMIR scattering so that
it works for OMIR files and for OMIRAnnotations.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>