[FIRRTL][InferReset] Keep instance graph properly updated (#1912)
In an earlier change, I made InferResets always mark the InstanceGraph
analysis as preserved, under the assumption that the pass does not
modify the instance graph.  This turns out to be incorrect as it can add
arguments to a module and must recreate the instance ops to reflect
that.

This change adds a function `InstanceGraph::replaceInstance(old, new)`
which allows the InstanceGraph to stay updated by the pass. This is the
first function added to the InstanceGraph to allow updating the cached
analysis.  This function is slightly more efficient then adding a
separate `addInstance(new)` and removeInstance(old)` functions.