[ExportVerilog] Remove module name mgmt from Prepare entirely.
This moves all the logic for this back into the main "CollectNames"
pass instead of doing some of it there and some of it in Prepare.
This has the benefit of making ModuleNameManager strictly a concern
for ExportVerilog, and makes sure that all the stuff generated by
the prepare pass is consistently noticed.