[FIRRTL] Expand coverage for printf-encoded verif ops (#1938)
Chisel and some of the SiFive-internal code make use of a variety of
asserts encoded as printf statements to get around limitations of the
native (possibly MFC-only) verification operations in FIRRTL. These are
generally encoded as `printf` in a `when` block:

    when somethingGoingWrong:
      printf(..., "assert:some_label:Oops, something wrong here")

The flavors include "assert", "assume", "cover", "assertNotX", but also
variations of "Assertion failed" and tags like "[verif-library-assert]".
There's also precedence for XML tags wrapping a JSON-serialized Scala
object describing additional properties of the assertion. Since we need
to reason about these in `ExtractTestCode` and when we generate Verilog,
these constructs should be converted to dedicated `firrtl.assert` ops
(and friends), which have clear semantics.

This change introduces a post-processing step for `when` blocks into the
FIRRTL parser, which pattern matches against these constructs and
replaces them with an actual `firrtl.{assert,assume,cover}` which
encodes the actual verification operation.