[FIRRTL] Add FIRRTL memory summary struct (#1970)
The `FirMemory` is a struct with the relevant `FIRRTL` memory properties. 
It was constructed during `LowerToHW` for lowering the `FIRRTL::MemOp` by 
extracting its attributes.
But the `FirMemory` is also required during the memory metadata generation.
In this checkin we move the `FirMemory` struct out of `LowerToHW` and add it 
to the `FIRRTLOps.h`. 
This also fixes the memory names emitted during the metadata generation.