[FIRRTL] Allow addToAlwaysBlock outside HWModuleOp
Resolve a comment and assert in `LowerToHW` that prevents
`addToAlwaysBlock` from working if it is nested withing anything except
for `HWModuleOp`. This commit adds the parent region as a key into the
map of constructed always blocks, such that arbitrary nesting is
allowed.

This was triggered by work on lowering of complex asserts, which require
nesting always blocks within ifdefs.