[FIRRTL] Mark extmodules that had black box anno consumed (#1996)
Fix an issue where generating metadata after the blackbox reader pass
has run would cause additional blackbox modules to appear in the
generated metadata. This happens because the blackbox reader removes the
blackbox annotations, which the metadata pass is looking for to
blacklist the corresponding extmodules.

This change causes the blackbox reader pass to add a `BlackBox`
annotation on any extmodule for which it has processed one of the
blackbox annotations. The metadata pass then additionally blacklists
this annotation. Also add a unit and integration test to ensure this
happens.

Fixes an issue in the SiFive flow.