[PyCDE] Set local_scheme to "no-local-version". (#1999)
This removes the local part of the version, which includes a commit
hash. The commit hash is redundant since the number of commits since a
version is already included in the version string, and we don't
rewrite git history. The local part of the version isn't compatible
with package registries, so we remove it.