[HW] Improve TypeAliasType::getTypeDecl.
This makes this more efficient (not doing a full mlir::Module walk to
look up uncached symbols), makes the diagnostics in ExportVerilog
come out with the correct location info again, and changes this accessor
to return null on failure, instead of emitting an error from the accessor.

There are still larger design changes that would make sense here
conceptually, but I'm not going to work on typedefs much.