[FIRRTL][GrandCentral] Make non-local paths local (#2040)
This is a follow on to 4d1c41e52084, where we adapted to a new path
format for Grand Central `AugmentedGroundType`s. The problem with this new
format is that all created annotation targets are non-local, which is
not applying correctly using NLAs. This change takes all given
non-local targets for augmented types and changes them to local targets,
without regard to whether these targets are equivalent.  This works for now
because all non-local targets and their corresponding local target
are in fact equivalent.

The proper change is to either a) attach these as non-local annotations
and add support for non-local annotations to Grand Central or b) have a
target canonicalizer which correctly makes these non-local annotations
local only when legal to do so.

Since this problem is currently blocking integration testing, I am
proposing this change for now.  I will be looking at moving the
GrandCentral annotation scattering into the newer
`firrtl-lower-annotations` pass, where hopefully we will accomplish
option `b`.