[ExportVerilog] Speed up emission of procedural decl declarations.  NFC.
It turns out that a huge chunk of ExportVerilog time is spent memcpy'ing
megabytes of text around to insert "begin" markers, and to insert the
declarations for procedural expressions that get spilled to a logic.

Optimize and simplify the code for this by emitting the declarations into
a temp buffer, then inserting that buffer into place.  This avoids one of
the copies of the large data.  This speeds up ExportVerilog on a large
testcase from 7.8 to 5.5s on my machine.

Note that this just _reduces_ the massive memcpy's, it doesn't fix them.
This patch makes way for a bigger change.