[ExportVerilog] Revamp module emission logic to not copy strings, NFC.
This introduces a new data structure "RearrangableOStream" which acts
like an ostream but with cursors and methods that allow moving around
blobs of text without copying the string data around.  This accumulates
the text data in a normal way, but has an additional segment list
imposed on top of it, that allows these efficient reorderings.

This speeds up ExportVerilog by about 8% on large testcases.  We
were copying around _megabytes_ of .sv text to insert begin statements
and "automatic logic" declarations, which isn't super fast.  On one
large example this sped it up from 4.75 to 4.4s.