[FIRRTL] Allow invalid values to be emitted as `is invalid`
Extend the `FIREmitter` to allow for `InvalidValueOp` uses in
expressions. Until now, the emitter would just error out if it saw an
invalid value used as an operand outside of a connect, for example as:

    %invalid_clock = firrtl.invalidvalue : !firrtl.clock
    %x = firrtl.reg %invalid_clock : !firrtl.uint<42>

This pattern can occur as a result of optimizations and transformations
done on the IR, though. This commit extends the `FIREmitter` to produce
a temporary wire for any invalid values that are used outside of
connects. In the above example, this now prints:

    wire _invalid : Clock
    _invalid is invalid
    reg x : UInt<42>, _invalid