[FIRRTL] Output GCT Taps into Extraction Directory (#2054)
Change the Grand Central Taps pass to output (via the output file
attribute) data and memory taps it creates to a directory indicated by
an optional ExtractGrandCentralAnnotation.  If this annotation does not
exist, do nothing.  This change is made to align Grand Central behavior
with the Scala FIRRTL Compiler.

Move the string definition of `ExtractGrandCentralAnnotation` into the
common Annotation header as this string is needed by all Grand Central
passes.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>