[Support] Use replaceAllUsesWith in BackedgeBuilder (#2079)
If a backedge value is referenced multiple times, for some reason, only the last reference is being replaced through the current `setValue` implementation. This is fixed by replacing the value using the more canonical approach of `Value::replaceAllUsesWith`.

Furthermore, adds a check to ensure that a backedge is only set one. I think this is a fair restriction, given that the backedge is a stand-in for an eventual SSA value.