[ExportVerilog] Block Comment for Bound Instance
Change emission of bound instances to use a block comment instead of an
actual comment applied per-line when handling the instance.  Per-line
emission causes problems because the output buffer may later be broken
to align with output emission options.  Instead, use a block comment so
that the output buffer can be reformatted in arbitrary ways without
having to pass around information that we're in a commented region.

Fixes #2092.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>