[FIRRTL] Add inner names to declarations (#2081)
Add a `inner_sym` attribute to the following FIRRTL dialect operations:

- `InstanceOp`
- `NodeOp`
- `WireOp`
- `RegOp`
- `RegResetOp`
- `MemOp`

The `inner_sym` attribute allows these operations to be named in a
`hw.innerNameRef`, and gives us the ability to attach a name to these
operations that will be carried over through lowering into the HW
dialect.

An interesting case arises if one of these operations has an inner name
and type lowering would have to duplicate the operation. This can occur
for example for a wire with aggregate type:

    %x = firrtl.wire sym @foo : !firrtl.bundle<a: uint<1>, b: uint<2>>

Type lowering would convert this into two wires, one for field `a` and
one for field `b`. However, since the wire itself defines the symbol
`@foo`, it is unclear what should happen with the symbol. At the moment,
type lowering will just emit an error in this case. Other valid options
would be to drop the symbol (which is kind of brittle), or introduce a
breadcrumb operation that keeps track of the change, like:

    %x_a = firrtl.wire sym @foo_a
    %x_b = firrtl.wire sym @foo_b
    firrtl.aggregate_symbol @foo, [@foo_a, @foo_b]

Since we have no such use cases at the moment, just emitting an error
seems fine.

This change does not yet include module port symbols, which require
significant additional changes that will follow as a separate commit.