[HW] Add support for port symbols to module builder (#2117)
Add a `sym` field to `PortInfo` which holds the optional `hw.exportPort`
symbol to use for a port. Extend the `buildModule` function to honor
this field and populate the `argAttrs` and `resultAttrs` arrays as
appropriate to contain these port symbols.

This will make it easier later for other dialects to attach port symbols
during the lowering to the HW dialect (FIRRTL is an example that is
coming up soon).