[FIRRTL] Make GCT data tap hierarchical names start at module (#2047)
Change how GCT data/memory taps are generated by always starting the
hierarchical names with a module. This removes ambiguities where a local
declaration `bar` could shadow an instance `bar` in a parent module,
rendering the XMR `bar.x` invalid. Also, if a signal is tapped in the
same module where the data taps module is instantiated, the resulting
hierarchical name would just be `x`, which in SV is not a hierarchical
name and causes no resolution against the `x` in the parent module we're
interested in. Always inserting a module name prefix removes any such
ambiguity, generates a hierarchical name in any case, and causes SV to
properly look upwards in the instance graph until that parent module is
found.