[MSFT] Add Tcl export support for PhysicalRegions. (#2155)
This is a very simple start to Tcl support for PhysicalRegions. For
now physical region references and the PlacedInstances making the
reference are just stored in a vector that can be walked to emit the
necessary Tcl. More complex storage and queries for PhysicalRegions
can be added as the need arises.