[Handshake] Ensure `argNames` and `resNames` attributes (#2144)
This commit ensures that the `argNames` and `resNames` attributes are always present in a `handshake.func` operation. The names are either partially (if some names were already provided to the op during conversion) or fully inferred based on whatever information available when parsing or building the `handshake.func` op.
In doing so, we move signal name generation from handshakeToFIRRTL into being part of the operation itself.