[FIRParser/ExportVerilog] Rework the change for Issue #2076
We want ExportVerilog to have predictable behavior no matter what
the input IR is.  The fix to that issue added some special behavior
for 0 values, but the actual problem is all the way up in the parser
StringRef::getAsInteger parser used by the firrtl frontend, which puts
zeros into an arbitrary 64-bit value.  Change that in the .fir parser
to properly confine integer constants to 32-bits when they fit.

This eliminates the firrtl specific logic from the verilog printer.