[HW][SV][ExportVerilog] Add Comment to HWModuleOp, Verilog Emission (#2066)
Add a new mandatory attribute, "comment", to HWModuleOp that describes a
string comment that should be attached to that module when this module
is emitted by a backend.

Add support for emitting comments associated with HWModuleOps during
ExportVerilog.  Use a simple line breaking algorithm to try to respect
the emittedLineLength option when reasonably possible.  Respect manual
line breaks that a user may have included in the comment.

Add tests that comments are emitted and that line breaking changes as
expected with changes to emittedLineLength.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>