[CombFolds] Don't fold SExtOp into a concat anymore.
I added this before with the understanding that a sext operation
is "just wires" so flattening it into a concat can unblock other
optimizations.  This is true, but doing so significantly pessimizes
generated verilog quality, and bloats out the IR in practice.

A better solution to this is to introduce a replicate op and remove
sextop entirely.  This will compose better with existing xforms.
I filed Issue #2248 to track this idea.