[PyCDE] Add support for placing CompRegOp. (#2304)
This extends the Instance wrapper to wrap not just instances but also
registers. Additionally, in the Tcl export, this looks for registers,
adds them to the symbol cache, and uses their symbol names.

Users are now able to place CompRegOp when its innerSym attribute is
set. If they create a CompRegOp directly, the inner_sym must be passed
explicitly, but the `reg(...)` helper on Value is extended to use the
name as the inner_sym when passed.