[ExportVerilog] Fix a super obscure bug where we didn't indent temps right.
In an `ifdef region, when spilling a temporary, we'd indent at the level
of the ifdef instead of the correct level of the definition, leading to
things like this (taken from the example in Issue #2310):

```
module Foo(
  input  clock, d,
  output q);

  reg r;

  `ifndef SYNTHESIS
    reg [31:0] _T;

  wire _T_0 = _T[0];   /// <--- indented wrong
    initial begin
```

The "unindent" behavior was needed when spilling due to lines too long,
because the decls are put outside the `ifdef region.  To solve this,
just keep track of the indent depth.