[MSFT] Design partition support, part 1 (#2366)
The "design partition" feature supports scenarios wherein the designer needs the RTL output module hierarchy to be different than the logical hierarchy they are designing. With this feature, they just create a design partition then "tag" various entities with it. Upon compilation, CIRCT will refactor the hierarchy to move the tagged entities into the specified target partition. The design partition _must_ be found in the module hierarchy above (or in the same module as) the tagged entities.

Partial progress towards #2365.

- Modeling in CIRCT and PyCDE
- Move entities in the same module as the partition