[MSFT] Refactor placement entry to use hw::GlobalRefs.
This replaces the use of msft::SwitchInstanceAttr on leaf entities
with a hw::GlobalRefOp. The global ref specifies the same path, using
an array of hw::InnerRefAttrs. It also has an attribute containing the
same placement information as the msft::SwitchInstanceAttr had.

This will simplify ExportQuartusTcl by letting it simply generate Tcl
referring to the hw::GlobalRefs. It will also make the placements
robust to changes in the instance hierarchy via the guarantees
provided by hw::GlobalRef's verifier and utilities for rebuilding
hw::GlobalRefs.