[LowerTypes][FIX] Make aggregate preservation more conservative
This commit makes aggregate preservation more conservative.
There are basically two changes: Previously bitcast was
fine to preserve aggregates but it breaks the compile if
`repl-seq-meme` is enabled. This issue is yet to be reduced but it
should not break the compile so I would like to make conservative
here. Another thing is the preservation of aggregate types with
zero-bit width(zero-bit integer or empty vector/bundle).
This is causing tedious downstream issues about zero-width struct
emissions in ExportVerilog so I would like to change not to emit
these things at all.