[SCFToCalyx] Initial support for PipelineWhileOp. (#2389)
This adds the necessary support for PipelineWhileOp in the
WhileOpInterface. It also adds basic support for lowering the
PipelineTerminatorOp and PipelineStageOp. Together, this enables the
PipelineWhileOp condition, iter args, and results to be converted.

Building pipeline registers and scheduling groups to pass values
between pipeline stages will be added as follow up work.