[FIRRTL] Add ParamDeclAttr, use for FExtModule (#2411)
Add HW Dialect's ParamDeclAttr to the FIRRTL Dialect.  This commit only
adds this attribute, but does not use it.  This commit is part of a
series of commits that migrate FIRRTL external modules to use HW-like
parameters.

Change FIRRTL's ExtModule to use ParamDeclAttr for parameter storage as
opposed to using a dictionary.  Modify parsing/printing to put
parameters in a carat-delimited area before ports (just like how HW
Dialect does this).

This makes NO modifications to the way ExtModule is represented.
Parameter values are still stored on the ExtModule and not on the
instantiation.  This will be changed in a later commit.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>