[StandardToHandshake] Initial commit for stateful std-to-hs lowering
This has been a long-standing to-do for me; the intention of this commit is to start removing lowering state from within the IR itself, and make it explicitly managed. In the long term, this intends to remove all of the "helper values" and assumptions on temporary values being placed at fixed indexes for certain operations - something which i don't see as maintainable nor extensible.

One of the first issues that is tackled through this is the managing of the "block entry control value". Previously, this was assumed to be any `control_merge` operation that existed within a basic block (an assumption which is no longer valid in some stuff i'm working on locally).