[CMake] Update header install commands to match MLIRs
Update the `install` commands concerned with the CIRCT header files to
match the MLIR style: one for `circt` and `circt-c` in the source tree,
and one for `circt` and `circt-c` in the build tree.

This only affects out-of-tree users of CIRCT and in the worst case
provide additional header files than what they have seen so far.