[SV] Add sv.system.sampled to model $sampled
Add SampledOp to represent the sampled value function $sampled.  This is
added to conform to SystemVerilog guidelines around how to emit
concurrent assertions which need to capture values in the true or false
action blocks of the assertion.  Section 16.9.3 of the SystemVerilog
specification goes into more detail here.

Add emission of sv.system.sampled as $sampled.  This is NOT
automatically used for the message ops of concurrent verification
statements and it is the responsibility of the user to properly use
these if they are trying to follow SystemVerilog best practices.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>