[FIRRTL] Update NonLocalAnchors in the  prefixModules pass (#2560)
This commit fixes the `NonLocalAnchor` after the modules are cloned.
PrefixModules copies all the nonlocal annotations when a module is cloned. 
Some of the NLAs can become invalid based on the instance path.
The modules are visited in a top down order based on the Instance graph.
The algorithm maintains an invariant that the instance path is always valid till the 
module which is being traversed. Hence, for any NLA anchor that exists in the module, 
we need to ensure, that the module occurs in the NLA path. 
Any NLA, that does not contain a reference to the current module should be removed.