[MSFT] Dynamic Instance op (#2622)
Represents an instance (as in instance in the instance hierarchy) referred
to henceforth as a dynamic instance. Specified with a path through the
instance hierarchy (which in the future will be replaced with an AppID).
Lowers to a `hw.globalref` but unlike a global ref, does not require all of
the ops participating in the globalref to contain a back pointer attribute.
Allows users to efficiently add placements to a large number of dynamic
instances which happen to map to a small number of static instances by
bulk-adding the necessary `hw.globalref` attributes.

During the lowering, moves the operations in the body to the top level and
gives them the symbol of the globalref which was created to replace the
dynamic instance.