[Calyx] Add attributes for external memory interfaces (#2649)
Currently doing some work on figuring out how to hook up Calyx with the cosimulation infrastructure in CIRCT-HLS. For this, i need to determine the set of ports which specifies a memory bundle. While this is possible via. a purely string-based approach, i'd like to be a bit safer and actually add some tagging information in SCFToCalyx to help resolve the ports that constitutes a memory interface.

This initial commit just adds the attributes to the component ports in SCFToCalyx, and does not intrude into the Calyx library itself.