[FIRParser] Flip assert predicate before applying modifiers
When parsing printf-encoded assertions, the condition of the surrounding
`when` has to be inverted in order to go from the fail-when-true
polarity of the printf to the fail-when-false polarity of the assert.
Predicate modifiers (currently just `trueOrIsX`) are formulated with
respect to the assertion predicate, i.e., on the fail-when-false
polarity. So a `trueOrIsX` modifier is supposed to add an additional
pass case to the assertion if the condition is `X`. The parser currently
applies predicate modifiers on the fail-when-true polarity, which is
wrong. Move the predicate inversion upwards so it occurs before
modifiers are applied.

Fixes #2653.