[FIRRTL] Support NLA Leaves in Grand Central  (#2632)
Fix a bug in Grand Central where it would not handle leaves which were
targeted using non-local annotations.  This bug shows up when using
Grand Central in combination with Deduplication.

Modify Grand Central to remove dead non-local annotations (NLAs) used to
mark leaves of XMRs.  This enables Grand Central to pass NLA
verification and not leave the circuit in an invalid state.

Add a test that exercises Grand Central with non-local
annotations (NLAs) used to mark leaves of XMRs.  This mimics the
situation where deduplication creates an NLA that Grand Central has to
deal with.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>