[FIRRTL][Dedup] Fix partial connect expansion (#2681)
When deduplicating two modules, the names of fields in bundle types may
change, which means we have to go and fix partial connect operations to
make sure that the proper fields remain connected.  There was a spelling
mistake in this piece of code which caused the expander to think that
the LHS and RHS types of the partial connect were unchanged, and not fix
up the partial connect.  This would typically cause fields which used to
be connected to become unconnected.