[FIRRTLToHW] Consider strictconnect for elimination (#2693)
The `LowerToHW` conversion uses the `tryEliminatingConnectsToValue`
function to try and eliminate unnecessary connects to a value and
replace all value uses with the connected value instead. This is used
to avoid unnecessary `*.output` temporary wires when generating the
`hw.output` lowering. That function only considers `ConnectOp` though,
and any `StrictConnectOp` are left untouched. This commit extends the
elimination to also consider `StrictConnectOp`, such that the lowered
HW design contains fewer temporary wires.